Move apply_transformer to kafka connector
